Move threaded scheduler tests into scheduler suite
