+++ tests/unit_test/pipeline/test_scheduler.py
@@ -2,8 +2,11 @@
 
 from __future__ import annotations
 
+import threading
+
 from sglang_omni_v1.scheduling.messages import IncomingMessage
 from sglang_omni_v1.scheduling.simple_scheduler import SimpleScheduler
+from sglang_omni_v1.scheduling.threaded_simple_scheduler import ThreadedSimpleScheduler
 from tests.unit_test.pipeline.helpers import run_scheduler
 
 
@@ -43,3 +46,54 @@ def test_simple_scheduler_batch_and_error_contracts() -> None:
     assert all(
         out.type == "error" and isinstance(out.data, ValueError) for out in outputs
     )
+
+
+def test_threaded_simple_scheduler_runs_requests_concurrently() -> None:
+    # Covers concurrent worker execution before result emission.
+    started: list[str] = []
+    lock = threading.Lock()
+    both_started = threading.Event()
+    release = threading.Event()
+
+    def compute(payload: str) -> str:
+        with lock:
+            started.append(payload)
+            if len(started) == 2:
+                both_started.set()
+        assert release.wait(timeout=2.0)
+        return payload
+
+    def wait_for_both_started() -> None:
+        try:
+            assert both_started.wait(timeout=2.0)
+        finally:
+            release.set()
+
+    outputs = run_scheduler(
+        ThreadedSimpleScheduler(compute, max_concurrency=2),
+        [
+            IncomingMessage("req-1", "new_request", "one"),
+            IncomingMessage("req-2", "new_request", "two"),
+        ],
+        output_count=2,
+        before_collect=wait_for_both_started,
+    )
+
+    assert {output.request_id for output in outputs} == {"req-1", "req-2"}
+    assert {output.data for output in outputs} == {"one", "two"}
+
+
+def test_threaded_simple_scheduler_reports_worker_errors() -> None:
+    # Covers worker exception emission as scheduler errors.
+    def compute(payload: str) -> str:
+        raise RuntimeError(payload)
+
+    outputs = run_scheduler(
+        ThreadedSimpleScheduler(compute, max_concurrency=1),
+        [IncomingMessage("req-err", "new_request", "boom")],
+        output_count=1,
+    )
+
+    assert outputs[0].request_id == "req-err"
+    assert outputs[0].type == "error"
+    assert isinstance(outputs[0].data, RuntimeError)
